refactor: colocate student name formatting
